+++ lib/api/loan/acalaApiLoan.dart
@@ -17,4 +17,27 @@ class AcalaApiLoan {
         .map((e) => LoanType.fromJson(Map<String, dynamic>.of(e)))
         .toList();
   }
+
+  Future<List<TotalCDPData>> queryTotalCDPs(List<String> pools) async {
+    final List res = await service.queryTotalCDPs(pools);
+    int index = 0;
+    return res.map((e) {
+      e['token'] = pools[index];
+      index++;
+      return TotalCDPData.fromJson(e);
+    }).toList();
+  }
+
+  Future<List<CollateralIncentiveData>> queryCollateralIncentives() async {
+    final res = await service.queryCollateralIncentives();
+    return res
+        .map((e) => CollateralIncentiveData.fromJson((e as List)))
+        .toList();
+  }
+
+  Future<List<CollateralRewardData>> queryCollateralRewards(
+      List<String> collaterals, String address) async {
+    final res = await service.queryCollateralRewards(collaterals, address);
+    return res.map((e) => CollateralRewardData.fromJson(e)).toList();
+  }
 }