+++ lib/api/types/homaRedeemAmountData.g.dart
@@ -9,10 +9,10 @@ part of 'homaRedeemAmountData.dart';
 HomaRedeemAmountData _$HomaRedeemAmountDataFromJson(Map<String, dynamic> json) {
   return HomaRedeemAmountData()
     ..atEra = json['atEra'] as int
-    ..amount = (json['amount'] as num)?.toDouble()
-    ..demand = (json['demand'] as num)?.toDouble()
-    ..fee = (json['fee'] as num)?.toDouble()
-    ..received = (json['received'] as num)?.toDouble();
+    ..amount = (json['amount'] as num).toDouble()
+    ..demand = (json['demand'] as num).toDouble()
+    ..fee = (json['fee'] as num).toDouble()
+    ..received = (json['received'] as num).toDouble();
 }
 
 Map<String, dynamic> _$HomaRedeemAmountDataToJson(