+++ lib/api/types/stakingPoolInfoData.g.dart
@@ -9,24 +9,20 @@ part of 'stakingPoolInfoData.dart';
 StakingPoolInfoData _$StakingPoolInfoDataFromJson(Map<String, dynamic> json) {
   return StakingPoolInfoData()
     ..rewardRate = json['rewardRate'] as String
-    ..freeList = (json['freeList'] as List)
-        ?.map((e) => e == null
-            ? null
-            : StakingPoolFreeItemData.fromJson(e as Map<String, dynamic>))
-        ?.toList()
-    ..claimFeeRatio = (json['claimFeeRatio'] as num)?.toDouble()
-    ..unbondingDuration = (json['unbondingDuration'] as num)?.toDouble()
-    ..freePool = (json['freePool'] as num)?.toDouble()
-    ..unbondingToFree = (json['unbondingToFree'] as num)?.toDouble()
+    ..freeList = (json['freeList'] as List<dynamic>)
+        .map((e) => StakingPoolFreeItemData.fromJson(e as Map<String, dynamic>))
+        .toList()
+    ..unbondingDuration = (json['unbondingDuration'] as num).toDouble()
+    ..freePool = (json['freePool'] as num).toDouble()
+    ..unbondingToFree = (json['unbondingToFree'] as num).toDouble()
     ..liquidTokenIssuance = json['liquidTokenIssuance'] as String
-    ..defaultExchangeRate = (json['defaultExchangeRate'] as num)?.toDouble()
-    ..maxClaimFee = (json['maxClaimFee'] as num)?.toDouble()
-    ..bondingDuration = (json['bondingDuration'] as num)?.toDouble()
-    ..currentEra = (json['currentEra'] as num)?.toDouble()
-    ..communalBonded = (json['communalBonded'] as num)?.toDouble()
-    ..communalTotal = (json['communalTotal'] as num)?.toDouble()
-    ..communalBondedRatio = (json['communalBondedRatio'] as num)?.toDouble()
-    ..liquidExchangeRate = (json['liquidExchangeRate'] as num)?.toDouble();
+    ..defaultExchangeRate = (json['defaultExchangeRate'] as num).toDouble()
+    ..bondingDuration = (json['bondingDuration'] as num).toDouble()
+    ..currentEra = (json['currentEra'] as num).toDouble()
+    ..communalBonded = (json['communalBonded'] as num).toDouble()
+    ..communalTotal = (json['communalTotal'] as num).toDouble()
+    ..communalBondedRatio = (json['communalBondedRatio'] as num).toDouble()
+    ..liquidExchangeRate = (json['liquidExchangeRate'] as num).toDouble();
 }
 
 Map<String, dynamic> _$StakingPoolInfoDataToJson(
@@ -34,13 +30,11 @@ Map<String, dynamic> _$StakingPoolInfoDataToJson(
     <String, dynamic>{
       'rewardRate': instance.rewardRate,
       'freeList': instance.freeList,
-      'claimFeeRatio': instance.claimFeeRatio,
       'unbondingDuration': instance.unbondingDuration,
       'freePool': instance.freePool,
       'unbondingToFree': instance.unbondingToFree,
       'liquidTokenIssuance': instance.liquidTokenIssuance,
       'defaultExchangeRate': instance.defaultExchangeRate,
-      'maxClaimFee': instance.maxClaimFee,
       'bondingDuration': instance.bondingDuration,
       'currentEra': instance.currentEra,
       'communalBonded': instance.communalBonded,
@@ -53,7 +47,7 @@ StakingPoolFreeItemData _$StakingPoolFreeItemDataFromJson(
     Map<String, dynamic> json) {
   return StakingPoolFreeItemData()
     ..era = json['era'] as int
-    ..free = (json['free'] as num)?.toDouble();
+    ..free = (json['free'] as num).toDouble();
 }
 
 Map<String, dynamic> _$StakingPoolFreeItemDataToJson(