+++ lib/service/serviceLoan.dart
@@ -1,5 +1,7 @@
 import 'package:polkawallet_plugin_acala/api/acalaApi.dart';
 import 'package:polkawallet_plugin_acala/api/types/loanType.dart';
+import 'package:polkawallet_plugin_acala/api/types/stakingPoolInfoData.dart';
+import 'package:polkawallet_plugin_acala/common/constants.dart';
 import 'package:polkawallet_plugin_acala/polkawallet_plugin_acala.dart';
 import 'package:polkawallet_plugin_acala/store/index.dart';
 import 'package:polkawallet_sdk/storage/keyring.dart';
@@ -25,6 +27,12 @@ class ServiceLoan {
         plugin.networkState.tokenDecimals[0]);
   }
 
+  Future<double> _fetchACAPrice() async {
+    final output =
+        await api.swap.queryTokenSwapAmount('1', null, ['ACA', 'AUSD'], '0.1');
+    return output.amount;
+  }
+
   Map<String, LoanData> _calcLoanData(
     List loans,
     List<LoanType> loanTypes,
@@ -44,11 +52,35 @@ class ServiceLoan {
     return data;
   }
 
+  Map<String, double> _calcCollateralIncentiveRate(
+      List<CollateralIncentiveData> incentives) {
+    final blockTime =
+        int.parse(plugin.networkConst['babe']['expectedBlockTime']);
+    final epoch =
+        int.parse(plugin.networkConst['incentives']['accumulatePeriod']);
+    final epochOfYear = SECONDS_OF_YEAR * 1000 / blockTime / epoch;
+    final res = Map<String, double>();
+    incentives.forEach((e) {
+      res[e.token] = Fmt.bigIntToDouble(
+              e.incentive, plugin.networkState.tokenDecimals[0]) *
+          epochOfYear;
+    });
+    return res;
+  }
+
   Future<void> queryLoanTypes(String address) async {
     if (address == null) return;
 
-    final loanTypes = await api.loan.queryLoanTypes();
-    store.loan.setLoanTypes(loanTypes);
+    final res = await Future.wait([
+      api.loan.queryLoanTypes(),
+      api.loan.queryCollateralIncentives(),
+    ]);
+    store.loan.setLoanTypes(res[0]);
+    if (res[1] != null) {
+      store.loan.setCollateralIncentives(_calcCollateralIncentiveRate(res[1]));
+    }
+
+    queryTotalCDPs();
   }
 
   Future<void> subscribeAccountLoans(String address) async {
@@ -59,12 +91,19 @@ class ServiceLoan {
     // 1. subscribe all token prices, callback triggers per 5s.
     api.assets.subscribeTokenPrices((Map<String, BigInt> prices) async {
       // 2. we need homa staking pool info to calculate price of LDOT
-      final stakingPoolInfo = await api.homa.queryHomaStakingPool();
+      final data = await Future.wait(
+          [api.homa.queryHomaStakingPool(), _fetchACAPrice()]);
+      final StakingPoolInfoData stakingPoolInfo = data[0];
       store.homa.setStakingPoolInfoData(stakingPoolInfo);
 
       // 3. set prices
       _calcLDOTPrice(prices, stakingPoolInfo.liquidExchangeRate);
+      prices['ACA'] =
+          Fmt.tokenInt(data[1].toString(), 18); // decimals of prices are 18.
       store.assets.setPrices(prices);
+
+      // 4. update collateral incentive rewards
+      queryCollateralRewards(address);
 
       // 4. we need loanTypes & prices to get account loans
       final loans = await api.loan.queryAccountLoans(address);
@@ -79,6 +118,18 @@ class ServiceLoan {
             _calcLoanData(loans, store.loan.loanTypes, prices));
       }
     });
+  }
+
+  Future<void> queryTotalCDPs() async {
+    final res = await api.loan
+        .queryTotalCDPs(store.loan.loanTypes.map((e) => e.token).toList());
+    store.loan.setTotalCDPs(res);
+  }
+
+  Future<void> queryCollateralRewards(String address) async {
+    final res = await api.loan.queryCollateralRewards(
+        store.loan.collateralIncentives.keys.toList(), address);
+    store.loan.setCollateralRewards(res);
   }
 
   void unsubscribeAccountLoans() {