+++ lib/store/loan.dart
@@ -17,7 +17,17 @@ abstract class _LoanStore with Store {
   List<LoanType> loanTypes = [];
 
   @observable
+  Map<String, TotalCDPData> totalCDPs = Map<String, TotalCDPData>();
+
+  @observable
   Map<String, LoanData> loans = Map<String, LoanData>();
+
+  @observable
+  Map<String, double> collateralIncentives = Map<String, double>();
+
+  @observable
+  Map<String, CollateralRewardData> collateralRewards =
+      Map<String, CollateralRewardData>();
 
   @observable
   bool loansLoading = true;
@@ -25,6 +35,29 @@ abstract class _LoanStore with Store {
   @action
   void setLoanTypes(List<LoanType> list) {
     loanTypes = list;
+  }
+
+  @action
+  void setTotalCDPs(List<TotalCDPData> list) {
+    final dataMap = Map<String, TotalCDPData>();
+    list.forEach((e) {
+      dataMap[e.token] = e;
+    });
+    totalCDPs = dataMap;
+  }
+
+  @action
+  void setCollateralIncentives(Map<String, double> data) {
+    collateralIncentives = data;
+  }
+
+  @action
+  void setCollateralRewards(List<CollateralRewardData> data) {
+    final dataMap = Map<String, CollateralRewardData>();
+    data.forEach((e) {
+      dataMap[e.token] = e;
+    });
+    collateralRewards = dataMap;
   }
 
   @action